+++ terraform/monitor.tf
@@ -28,6 +28,7 @@ resource "azurerm_monitor_log_profile" "j1dev_log_profile" {
   ]
 
   locations = [
+    "eastus",
     "westus",
     "global",
   ]
@@ -39,5 +40,16 @@ resource "azurerm_monitor_log_profile" "j1dev_log_profile" {
   retention_policy {
     enabled = true
     days    = 365
+  }
+}
+
+resource "azurerm_monitor_activity_log_alert" "j1dev" {
+  name                = "j1dev"
+  resource_group_name = azurerm_resource_group.j1dev.name
+  scopes              = [data.azurerm_subscription.j1dev.id]
+
+  criteria {
+    operation_name = "Microsoft.Authorization/policyAssignments/write"
+    category       = "Administrative"
   }
 }